+++ app/templates/src/app/actions/AppActions.js
@@ -1,20 +1,53 @@
-import AppDispatcher from '../dispatcher/AppDispatcher';
-import AppConstants from '../constants/AppConstants';
+import ActionTypes from '../constants/ActionTypes';
 // import WebAPIUtils from '../utils/WebAPIUtils';
 
-var AppActions = {
+/**
+ * App actions
+ *
+ * @type {Object}
+ */
+const AppActions = {
+    /**
+     * App initialize action
+     *
+     * @return {Object}
+     */
     initialize() {
-        AppDispatcher.dispatch({
-            type: AppConstants.INITIALIZE,
-            data: null
-        });
+        return {
+            type: ActionTypes.INITIALIZE
+        };
     },
 
-    placeholder(data) {
-        AppDispatcher.dispatch({
-            type: AppConstants.PLACEHOLDER,
-            data: data
-        });
+    /**
+     * Placeholder action (example)
+     *
+     * @param  {number} inc - increment value
+     *
+     * @return {Object}
+     */
+    placeholder(inc) {
+        return {
+            type: ActionTypes.PLACEHOLDER,
+            inc
+        };
+    },
+
+    /**
+     * Example of async action
+     *
+     * @param  {number} inc - increment value
+     *
+     * @return {Function}
+     */
+    placeholderAsync(inc) {
+        return function (dispatch) {
+            return new Promise((resolve) => {
+                setTimeout(() => {
+                    dispatch(AppActions.placeholder(inc));
+                    resolve();
+                }, 500);
+            });
+        };
     }
 };