+++ app/templates/src/app/constants/ActionTypes.js
@@ -1,0 +1,11 @@
+/**
+ * Action types constants
+ *
+ * @type {Object}
+ */
+const ActionTypes = {
+    INITIALIZE: 'INITIALIZE',
+    PLACEHOLDER: 'PLACEHOLDER'
+};
+
+export default ActionTypes;