+++ app/templates/src/app/constants/InitialState.js
@@ -1,0 +1,10 @@
+/**
+ * Initial states for reducers
+ *
+ * @type {Object}
+ */
+const InitialState = {
+    template: 0
+};
+
+export default InitialState;