+++ app/templates/src/app/reducers/index.js
@@ -1,0 +1,14 @@
+import { combineReducers } from 'redux';
+import { routerReducer as routing } from 'react-router-redux';
+
+import template from './templateReducer';
+
+/**
+ * Combine reducers
+ */
+const rootReducer = combineReducers({
+    template,
+    routing
+});
+
+export default rootReducer;