+++ app/templates/src/app/routes.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import { Route, Redirect } from 'react-router';
 
-import Template from './components/Template/Template.react';
+import Template from './components/Template/Template';
 
+/**
+ * Define routes
+ */
 var routes = (
     <Route path="/" component={Template}>
         <Redirect from="*" to="/" />