+++ app/templates/src/entry/main.js
@@ -1,6 +1,6 @@
 import 'babel-polyfill';
 import 'modernizr';
-import 'normalize.css/normalize.css';
+import 'normalize.css';
 
 // import global styles
 import './style/global/global.scss';
@@ -9,23 +9,23 @@ import './style/global/global.scss';
 import copy from '../copy/copy.json';
 
 // render incompatible browser screen
-import incompatible from './utilities/incompatible';
+import Incompatible from './utilities/Incompatible';
 import incompatibleTemplate from './views/incompatible/incompatible.html';
 import incompatibleStyle from './views/incompatible/incompatible.scss';
 
-incompatible.addClass();
-incompatible.render(incompatibleTemplate, incompatibleStyle, copy.incompatible);
+Incompatible.addClass();
+Incompatible.render(incompatibleTemplate, incompatibleStyle, copy.incompatible);
 
 // render build version
-import version from './utilities/version';
+import Version from './utilities/Version';
 import versionTemplate from './views/version/version.html';
 import versionStyle from './views/version/version.scss';
 
-version.logConsole();
-version.render(versionTemplate, versionStyle, version.getCopy());
+Version.logConsole();
+Version.render(versionTemplate, versionStyle, Version.getCopy());
 
 // render loader
-import loader from './utilities/loader';
+import Loader from './utilities/Loader';
 import loaderTemplate from './views/loader/loader.html';
 import loaderStyle from './views/loader/loader.scss';
 
@@ -34,7 +34,7 @@ function _progress(e) {
     var p = Math.round(100 * e.progress);
 
     // render progress in loader
-    loader.render(loaderTemplate, loaderStyle, { loader: p + copy.loader.progress });
+    Loader.render(loaderTemplate, loaderStyle, { loader: p + copy.loader.progress });
 }
 
 function _complete() {
@@ -43,7 +43,7 @@ function _complete() {
         var App = require('../app/App').default;
 
         // hide loader
-        loader.hide();
+        Loader.hide();
 
         // run app
         App.run();
@@ -52,12 +52,12 @@ function _complete() {
 
 // bootstrap application
 window.addEventListener('load', () => {
-    if (!incompatible.isIncompatibleBrowser()) {
+    if (!Incompatible.isIncompatibleBrowser()) {
         // show loader
-        loader.render(loaderTemplate, loaderStyle, { loader: copy.loader.start });
-        loader.show();
+        Loader.render(loaderTemplate, loaderStyle, { loader: copy.loader.start });
+        Loader.show();
 
         // start loader
-        loader.createLoader('main', _progress, _complete);
+        Loader.createLoader('main', _progress, _complete);
     }
 });