+++ app/templates/webpack.dev.config.js
@@ -4,11 +4,17 @@ import webpackConfig from './webpack.config.js';
 
 var config = extend(true, {}, webpackConfig, {
     entry: [
-        'webpack-hot-middleware/client?reload=true&quiet=true',
+        'webpack-hot-middleware/client?reload=true',
         './src/entry/main.js'
     ],
 
     devtool: '#cheap-module-source-map',
+
+    module: {
+        loaders: webpackConfig.module.loaders.concat([
+            { test: /\.js$/, exclude: /node_modules/, loader: 'react-hot!babel' }
+        ])
+    },
 
     plugins: webpackConfig.plugins.concat([
         new webpack.DefinePlugin({