+++ app/templates/webpack.prod.config.js
@@ -5,6 +5,14 @@ import webpackConfig from './webpack.config.js';
 var config = extend(true, {}, webpackConfig, {
     entry: './src/entry/main.js',
 
+    devtool: '#source-map',
+
+    module: {
+        loaders: webpackConfig.module.loaders.concat([
+            { test: /\.js$/, exclude: /node_modules/, loader: 'babel' }
+        ])
+    },
+
     plugins: webpackConfig.plugins.concat([
         new webpack.DefinePlugin({
             'process.env': { NODE_ENV: JSON.stringify('production') }